Extracting ui hooks
