+++ apps/rays-dashboard/components/molecules/BridgeSwap/BridgeSwapOnboarding.tsx
@@ -1,9 +1,8 @@
-import { Button, EXTERNAL_LINKS, Text } from '@summerfi/app-ui'
+import { Button, EXTERNAL_LINKS, Text, useOnboarding } from '@summerfi/app-ui'
 import Image from 'next/image'
 import Link from 'next/link'
 
 import { basePath } from '@/helpers/base-path'
-import { useOnboarding } from '@/helpers/use-onboarding'
 
 export const BridgeSwapOnboarding = () => {
   const [, setAsOnboarded] = useOnboarding('SwapWidget')
