+++ packages/app-ui/src/hooks/use-toggle.ts
@@ -1,8 +1,9 @@
+'use client'
 import { type Dispatch, type SetStateAction, useCallback, useState } from 'react'
 
-export function useToggle(
+export const useToggle = (
   initialState: boolean,
-): [boolean, () => void, Dispatch<SetStateAction<boolean>>] {
+): [boolean, () => void, Dispatch<SetStateAction<boolean>>] => {
   const [state, setState] = useState<boolean>(initialState)
 
   const toggle = useCallback(() => setState((nextState) => !nextState), [])
