+++ packages/app-ui/src/index.ts
@@ -40,7 +40,11 @@ export { Navigation } from './components/layout/Navigation/Navigation'
 export { NavigationMenuDropdownContentListItem } from './components/layout/Navigation/NavigationMenuDropdownContentListItem'
 
 export { EXTERNAL_LINKS, INTERNAL_LINKS } from './helpers/application-links'
-export { useClientSideMount } from './helpers/use-client-side-mount'
+
+export { useClientSideMount } from './hooks/use-client-side-mount'
+export { useLocalStorage, getStorageValue } from './hooks/use-local-storage'
+export { useToggle } from './hooks/use-toggle'
+export { useOnboarding } from './hooks/use-onboarding'
 
 // tokens
 export { tokenConfigs } from './tokens/config'
