+++ packages/app-utils/README.MD
@@ -1,10 +1,3 @@
-# @summerfi/app-tos
+# @summerfi/app-utils
 
-Common Terms of Service (TOS) handling for the @summerfi apps.
-
-This package provides `useTermsOfService` hook and types which can be used to easily integrate TOS
-flow in new apps.
-
-This package do not contain any UI which should be handled / customized on the client app side.
-
-This package assumes that the API already exists, which in the end may vary per specific app.
+Common utils for all apps
